requirements.txt
llama-index-core
llama-parse
llama-index-llms-gemini
python-dotenv
pydantic<2.10
llama-index-graph-stores-neo4j
